requirements.txt
flask==3.1.0
google-cloud-storage==2.19.0